Some corrections and minor changes.
Added semicolons where it looked like there should be semicolons.
Hopefully these were typos and not a needed omission. Changed
validateId to validateUserId to prevent confusion in the likely case
that similar functions are implemented in the other controllers.
Finally, added code to log the status code when using get.